moves formatISODate inside utils/date.ts
