+++ src/lib/utils/date.ts
@@ -39,3 +39,18 @@ export const getCurrentDateInPST = () => {
   const pstDate = DateTime.now().setZone('America/Los_Angeles').toISO();
   return pstDate;
 };
+
+export function formatISODate(isoDateString: string): string {
+  const date = new Date(isoDateString);
+
+  // Check if the date is valid
+  if (isNaN(date.getTime())) {
+      throw new Error('Invalid ISO date string');
+  }
+
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  const year = date.getFullYear();
+
+  return `${month}/${day}/${year}`;
+}
